Add basic highlighting to transcript segment
This commit adds logic for highlighting elements in the transcript.

In future this might have a variety of uses, but for now it is used to highlight entities.

This approach has a limitation that replacement instructions should be ordered from lowest start offset to highest.
Additionally, substrings identified by start and end offsets cannot overlap.

The replacement is applied from right to left. This avoids the problem of changing the string offset, which would occur if the replacements were applied from left to right.

I've not codified these assertions into the function itself

I wanted to keep this PR small, so all entities are highlighted red. In a future PR they will be highlighted different colours according to type.